api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api"
version = "0.0.1"
description = "API for Lunarr"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.58.0",
    "mem0ai==1.0.2",
    "pydantic>=2.12,<2.13",
    "pydantic-settings>=2.12,<2.13",
    "sentence-transformers==5.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9,<10",
    "pytest-asyncio==1.3.0",
    "ruff>=0.14,<0.15",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["lunarr"]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "BLE",      # flake8-blind-except
    "D",        # pydocstyle
    "EM",       # flake8-errmsg
    "FBT",      # flake8-boolean-trap
    "FIX",      # flake8-fixme
    "TC",       # flake8-type-checking
    "TD",       # flake8-todos
    "TRY",      # tryceratops
    "SIM108",   # if-else-block-instead-of-if-exp
    "COM812",   # missing-trailing-comma (due to conflicts with formatter)
    "E501",     # line-too-long
    "T201",     # print-statement
    "PLR2004",  # Magic value used in comparison
]

[tool.ruff.lint.pylint]
max-args = 15

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.per-file-ignores]
"**/tests/**" = [
    "ANN",      # flake8-annotations
    "S",        # flake8-bandit
    "SLF",      # flake8-self
    "ARG001",
]
"**/config.py" = [
    "S105",     # Possible hardcoded password (dev defaults)
]
